DPT-1007 Create CallableSchema interface to generalise function/procedure/etc
